+++ lib/Extension/LanguageServerIndexer/Listener/IndexerListener.php
@@ -31,9 +31,11 @@ class IndexerListener implements ListenerProviderInterface
         }
 
         if ($event instanceof WillShutdown) {
-            if ($this->manager->isRunning(IndexerHandler::SERVICE_INDEXER)) {
-                $this->manager->stop(IndexerHandler::SERVICE_INDEXER);
-            }
+            yield function (): void {
+                if ($this->manager->isRunning(IndexerHandler::SERVICE_INDEXER)) {
+                    $this->manager->stop(IndexerHandler::SERVICE_INDEXER);
+                }
+            };
         }
     }
 }